requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.3
passlib==1.8.1
python-jose==3.3.0
